Finish typing @vue/common
