+++ packages/awtrix/templates/app/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./node_modules/@awtrix/common/tsconfig.app.json"
+}
