+++ packages/awtrix/templates/core/apps/awtrix/0.0.1/backend.js
@@ -1,15 +1,5 @@
-const { Yeelight } = require('yeelight-node')
-
 module.exports = (BaseClass) => {
   return class extends BaseClass {
-    register () {
-      const light = new Yeelight({ ip: '10.0.0.51', port: 55443 })
-
-      this.io.on('connect', (client) => {
-        client.on('toggle', (ctx) => {
-          light.toggle()
-        })
-      })
-    }
+    register () { }
   }
 }
