+++ packages/common/src/index.ts
@@ -1,18 +1,4 @@
-export { GeneratorType, default as createFrontend } from './createFrontend'
+export { GeneratorType, default as defineComponent } from './defineComponent'
 export { default as BackendApp } from './BackendApp'
 export { default as FrontendApp } from './FrontendApp'
-import { foobar } from './createFrontend'
 export * as Types from './types'
-
-foobar({
-  data () {
-    return {
-      nuddel: 'foa'
-    }
-  },
-
-  created () {
-    this.nuddel
-    this.lock()
-  }
-})
