+++ packages/common/tsconfig.app.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+      "target": "esnext",
+      "module": "esnext",
+      "moduleResolution": "node",
+      "strict": true
+  }
+}
